refactor: Reuse modal manager hook in ModalRootDesktop
