+++ src/components/ModalRoot/ModalRootDesktop.tsx
@@ -11,13 +11,14 @@ import {
   ConfigProviderContextInterface,
   WebviewType,
 } from '../ConfigProvider/ConfigProviderContext';
-import { ModalsStateEntry, ModalType } from './types';
+import { ModalsStateEntry } from './types';
 import { ANDROID, VKCOM } from '../../lib/platform';
 import { getClassName } from '../../helpers/getClassName';
 import { DOMProps, withDOM } from '../../lib/dom';
 import { getNavId } from '../../lib/getNavId';
 import { warnOnce } from '../../lib/warnOnce';
 import { FocusTrap } from '../FocusTrap/FocusTrap';
+import { useModalManager, ModalTransitionProps } from './useModalManager';
 import './ModalRoot.css';
 
 const warn = warnOnce('ModalRoot');
@@ -36,40 +37,11 @@ export interface ModalRootProps extends HasPlatform {
   onClose?(modalId: string): void;
 }
 
-interface ModalRootState {
-  activeModal?: string;
-  prevModal?: string;
-  nextModal?: string;
-  visibleModals?: string[];
-  animated?: boolean;
-  switching?: boolean;
-  history?: string[];
-  isBack?: boolean;
-  inited?: boolean;
-}
-
-class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProps, ModalRootState> {
-  constructor(props: ModalRootProps) {
+class ModalRootDesktopComponentDumb extends React.Component<ModalRootProps & DOMProps & ModalTransitionProps> {
+  constructor(props: ModalRootProps & ModalTransitionProps) {
     super(props);
 
-    const activeModal = props.activeModal;
-
-    this.state = {
-      activeModal: null,
-      prevModal: null,
-      nextModal: activeModal,
-      visibleModals: activeModal ? [activeModal] : [],
-      animated: !!activeModal,
-      switching: false,
-      history: activeModal ? [activeModal] : [],
-      isBack: false,
-      inited: false,
-    };
-
     this.maskElementRef = React.createRef();
-    this.activeTransitions = 0;
-
-    this.initModalsState();
 
     this.modalRootContext = {
       updateModalHeight: () => undefined,
@@ -79,13 +51,10 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
     };
   }
 
-  private modalsState: { [id: string]: ModalsStateEntry };
   private readonly maskElementRef: React.RefObject<HTMLDivElement>;
   private maskAnimationFrame: number;
   private readonly modalRootContext: ModalRootContextInterface;
   private restoreFocusTo: HTMLElement;
-
-  activeTransitions: number;
 
   get timeout() {
     return this.props.platform === ANDROID || this.props.platform === VKCOM ? 320 : 400;
@@ -95,113 +64,49 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
     return React.Children.toArray(this.props.children) as React.ReactElement[];
   }
 
-  initModalsState() {
-    this.modalsState = this.modals.reduce<{ [id: string]: ModalsStateEntry }>((acc, Modal) => {
-      const modalProps = Modal.props;
-      const state: ModalsStateEntry = {
-        id: getNavId(Modal.props, warn),
-        onClose: Modal.props.onClose,
-        dynamicContentHeight: !!modalProps.dynamicContentHeight,
-      };
-
-      // ModalPage props
-      if (typeof modalProps.settlingHeight === 'number') {
-        state.settlingHeight = modalProps.settlingHeight;
-      }
-
-      acc[state.id] = state;
-      return acc;
-    }, {});
+  get modalsState() {
+    return this.props.modalsState;
   }
 
-  componentDidMount() {
-    this.initActiveModal();
-  }
+  componentDidUpdate(prevProps: ModalRootProps & ModalTransitionProps) {
+    // transition phase 2: animate exiting modal
+    if (this.props.exitingModal && this.props.exitingModal !== prevProps.exitingModal) {
+      this.closeModal(this.props.exitingModal);
+    }
 
-  componentDidUpdate(prevProps: ModalRootProps, prevState: ModalRootState) {
-    if (this.props.activeModal !== prevProps.activeModal) {
-      const nextModal = this.props.activeModal;
-      const prevModal = prevProps.activeModal;
-
-      if (IS_DEV && nextModal !== null && !this.modalsState[nextModal]) {
-        return warn(`[ModalRoot.componentDidUpdate] nextModal ${nextModal} not found`);
-      }
-
-      let history = [...this.state.history];
-      let isBack = false;
-
-      if (nextModal === null) {
-        history = [];
-      } else if (history.includes(nextModal)) {
-        history = history.splice(0, history.indexOf(nextModal) + 1);
-        isBack = true;
-      } else {
-        history.push(nextModal);
-      }
-
-      return this.setState({
-        activeModal: null,
-        nextModal,
-        prevModal,
-        visibleModals: [nextModal, prevModal],
-        history,
-        isBack,
-        animated: true,
-        inited: false,
-        switching: false,
-      }, () => {
-        if (nextModal === null) {
-          this.closeActiveModal();
-        } else {
-          this.initActiveModal();
+    // transition phase 3: animate entering modal
+    if (this.props.enteringModal && this.props.enteringModal !== prevProps.enteringModal) {
+      const { enteringModal } = this.props;
+      const enteringState = this.modalsState[enteringModal];
+      requestAnimationFrame(() => {
+        if (this.props.enteringModal === enteringModal) {
+          this.waitTransitionFinish(enteringState, () => this.props.onEnter(enteringModal));
+          this.animateModalOpacity(enteringState, true);
         }
       });
     }
 
-    if (this.state.switching && !prevState.switching) {
-      requestAnimationFrame(() => this.switchPrevNext());
+    // focus restoration
+    if (this.props.activeModal && !prevProps.activeModal) {
+      this.restoreFocusTo = this.props.document.activeElement as HTMLElement;
+    }
+    if (!this.props.activeModal && !this.props.exitingModal && this.restoreFocusTo) {
+      this.restoreFocusTo.focus();
+      this.restoreFocusTo = null;
     }
   }
 
-  /**
-   * Инициализирует модалку перед анимацией открытия
-   */
-  initActiveModal() {
-    const activeModal = this.state.activeModal || this.state.nextModal;
-    if (!activeModal) {
+  closeModal(id: string) {
+    const prevModalState = this.modalsState[id];
+    if (!prevModalState) {
       return;
     }
 
-    const modalState = this.modalsState[activeModal];
-
-    switch (modalState.type) {
-      case ModalType.PAGE:
-        modalState.settlingHeight = modalState.settlingHeight || 75;
-        break;
-
-      case ModalType.CARD:
-        break;
-
-      default:
-        if (IS_DEV) {
-          warn('[initActiveModal] modalState.type is unknown');
-        }
-    }
-
-    this.setState({ inited: true, switching: true });
-  }
-
-  closeActiveModal() {
-    const { prevModal } = this.state;
-    if (IS_DEV && !prevModal) {
-      return warn(`[closeActiveModal] prevModal is ${prevModal}`);
-    }
-
-    const prevModalState = this.modalsState[prevModal];
-
-    this.waitTransitionFinish(prevModalState, this.prevNextSwitchEndHandler);
+    this.waitTransitionFinish(prevModalState, () => this.props.onExit(id));
     this.animateModalOpacity(prevModalState, false);
-    this.setMaskOpacity(prevModalState, 0);
+    if (!this.props.activeModal) {
+      this.setMaskOpacity(prevModalState, 0);
+    }
   }
 
   waitTransitionFinish(modalState: ModalsStateEntry, eventHandler: () => void) {
@@ -217,82 +122,6 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
     }
   }
 
-  switchPrevNext() {
-    const { prevModal, nextModal } = this.state;
-
-    const prevModalState = this.modalsState[prevModal];
-    const nextModalState = this.modalsState[nextModal];
-
-    if (IS_DEV && !prevModalState && !nextModalState) {
-      return warn(`[switchPrevNext] prevModal is ${prevModal}, nextModal is ${nextModal}`);
-    }
-
-    const prevIsCard = !!prevModalState && prevModalState.type === ModalType.CARD;
-
-    const nextIsPage = !!nextModalState && nextModalState.type === ModalType.PAGE;
-    const nextIsCard = !!nextModalState && nextModalState.type === ModalType.CARD;
-
-    // Ждём полного скрытия предыдущей модалки
-    if (prevModalState && (nextIsCard || prevIsCard && nextIsPage)) {
-      this.activeTransitions += 1;
-      this.waitTransitionFinish(prevModalState, () => {
-        this.waitTransitionFinish(nextModalState, this.prevNextSwitchEndHandler);
-        this.animateModalOpacity(nextModalState, true);
-      });
-
-      requestAnimationFrame(() => {
-        this.animateModalOpacity(prevModalState, false);
-      });
-
-      return;
-    }
-
-    if (prevModalState && nextIsPage) {
-      this.activeTransitions += 1;
-      this.waitTransitionFinish(prevModalState, this.prevNextSwitchEndHandler);
-      requestAnimationFrame(() => {
-        this.animateModalOpacity(prevModalState, false);
-      });
-    }
-
-    if (!prevModal && this.props.document.activeElement) {
-      this.restoreFocusTo = this.props.document.activeElement as HTMLElement;
-    }
-
-    this.activeTransitions += 1;
-    this.waitTransitionFinish(nextModalState, this.prevNextSwitchEndHandler);
-    requestAnimationFrame(() => {
-      this.animateModalOpacity(nextModalState, true);
-    });
-  }
-
-  prevNextSwitchEndHandler = () => {
-    this.activeTransitions = Math.max(0, this.activeTransitions - 1);
-    if (this.activeTransitions > 0) {
-      return;
-    }
-
-    const activeModal = this.state.nextModal;
-
-    const newState: ModalRootState = {
-      prevModal: null,
-      nextModal: null,
-      visibleModals: [activeModal],
-      activeModal: activeModal,
-      animated: false,
-      switching: false,
-    };
-
-    if (!activeModal) {
-      newState.history = [];
-      if (this.restoreFocusTo) {
-        this.restoreFocusTo.focus();
-      }
-    }
-
-    this.setState(newState);
-  };
-
   /* Анимирует сдивг модалки */
   animateModalOpacity(modalState: ModalsStateEntry, display: boolean) {
     modalState.innerElement.style.opacity = display ? '1' : '0';
@@ -300,7 +129,7 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
 
   /* Устанавливает прозрачность для полупрозрачной подложки */
   setMaskOpacity(modalState: ModalsStateEntry, forceOpacity: number = null) {
-    if (forceOpacity === null && this.state.history[0] !== modalState.id) {
+    if (forceOpacity === null && this.props.history[0] !== modalState.id) {
       return;
     }
 
@@ -319,7 +148,7 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
    * Закрывает текущую модалку
    */
   triggerActiveModalClose = () => {
-    const activeModalState = this.modalsState[this.state.activeModal];
+    const activeModalState = this.modalsState[this.props.activeModal];
     if (activeModalState) {
       this.doCloseModal(activeModalState);
     }
@@ -336,9 +165,9 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
   };
 
   render() {
-    const { prevModal, activeModal, nextModal, visibleModals, animated } = this.state;
+    const { exitingModal, activeModal, enteringModal } = this.props;
 
-    if (!activeModal && !prevModal && !nextModal && !animated) {
+    if (!activeModal && !exitingModal) {
       return null;
     }
 
@@ -357,7 +186,7 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
           <div vkuiClass="ModalRoot__viewport">
             {this.modals.map((Modal: React.ReactElement) => {
               const modalId = getNavId(Modal.props, warn);
-              if (!visibleModals.includes(modalId)) {
+              if (modalId !== activeModal && modalId !== exitingModal) {
                 return null;
               }
 
@@ -370,9 +199,9 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
                   timeout={this.timeout}
                   key={key}
                   vkuiClass={classNames('ModalRoot__modal', {
-                    'ModalRoot__modal--active': modalId === activeModal,
-                    'ModalRoot__modal--prev': modalId === prevModal,
-                    'ModalRoot__modal--next': modalId === nextModal,
+                    'ModalRoot__modal--active': !exitingModal && !enteringModal && modalId === activeModal,
+                    'ModalRoot__modal--prev': modalId === exitingModal,
+                    'ModalRoot__modal--next': exitingModal && modalId === activeModal,
                   })}
                 >{Modal}</FocusTrap>
               );
@@ -383,5 +212,10 @@ class ModalRootDesktopComponent extends React.Component<ModalRootProps & DOMProp
     );
   }
 }
+
+const ModalRootDesktopComponent: React.FC<ModalRootProps> = (props) => {
+  const transitionManager = useModalManager(props.activeModal, props.children as any);
+  return <ModalRootDesktopComponentDumb {...props} {...transitionManager} />;
+};
 
 export const ModalRootDesktop = withContext(withPlatform(withDOM<ModalRootProps>(ModalRootDesktopComponent)), ConfigProviderContext, 'configProvider');
